Tribo : mobile layout, server side caching of games, notifications.
Games are now cached server side instead of being loaded from DB and computed each time a move is done.
Layout of the game is different for thin screens.
Moves are seen as new messages for observers and as pings for the other player (the tab title and icon change, and a desktop notification may appear)