small improvements of watches
If the mouse doesn't stay at least 250ms on the watch
it doesn't open and the unseen message count isn't
resetted.
The message count stays until the watch is closed.
Unseen messages are displayed over a yellow background
(theme dependent).
The number of displayed messages depends on the number
of unseen messages : more messages can be fetched, up
to 20 messages).