Tables looking like code are seen as tables
Example: "    A    |     B    \n-----+----\n    a    |    b"
This table is produced by postgresql. The spaces used to center
A make it look like a code bloc.
To solve this, we now look at the 3 first lines of a code
bloc and we decide it's a table when it looks like a table.