bulruk
The *bulruk* is a bubble displayed when hovering the *reply* mark
of a message when it replies to a not visible one. It shows
the content so that you don't have to click the mark.

With this commit also comes a new formating of the messages
in the watch drawers.