fix external-link-previewer breaking miaou
The out-of-memory error was happening when miaou was trying to "preview"
links to files of several GB. The fact miaou was also trying to parse it
as HTML wasn't helping either.

Now there's a limit on the accepted body size, and the mime type is also
checked.

The old request package had to be replaced by something a little more
modern and powerful. I took node-fetch.